removed deadlock

+++ philo/main.c
@@ -6,13 +6,14 @@
 /*   By: amalecki <amalecki@students.42wolfsburg    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/29 11:34:03 by amalecki          #+#    #+#             */
-/*   Updated: 2022/01/03 12:25:25 by amalecki         ###   ########.fr       */
+/*   Updated: 2022/01/03 13:00:44 by amalecki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-static void	init_philosophers(t_philo *p, pthread_mutex_t *forks, int args[6])
+static void	init_philosophers(t_philo *p, pthread_mutex_t *forks,
+	pthread_mutex_t *death_checker_locks, int args[6])
 {
 	int				i;
 
@@ -27,8 +28,18 @@ static void	init_philosophers(t_philo *p, pthread_mutex_t *forks, int args[6])
 		p[i].tsleep = args[3];
 		p[i].meals = args[4];
 		p[i].number = i + 1;
-		p[i].first_fork = &forks[i];
-		p[i].second_fork = &forks[(i + 1) % args[0]];
+		if (i == args[0] - 1)
+		{
+			p[i].first_fork = &forks[0];
+			p[i].second_fork = &forks[args[0] - 1];
+		}
+		else
+		{
+			p[i].first_fork = &forks[i];
+			p[i].second_fork = &forks[i + 1];
+		}
+		// p[i].first_fork = &forks[i];
+		// p[i].second_fork = &forks[(i + 1) % args[0]];
 		i++;
 	}
 }
@@ -68,13 +79,15 @@ int	main(int argc, char *argv[])
 	if (!aloc_pointers(&forks, &philosophers, &threads, args[0]))
 		return (write(1, "Problem with memory allocation!\n", 33));
 	init_forks(forks, args[0]);
-	init_death_checker_locks(&death_checker_locks, args[0]);
-	init_philosophers(philosophers, forks, args);
-	start_threads(threads, philosophers, args[0]);
-	start_death_checker_threads(threads, philosophers, args[0]);
-	join_threads(threads, args[0]);
+	if (init_death_checker_locks(&death_checker_locks, args[0]))
+	{
+		init_philosophers(philosophers, forks, death_checker_locks, args);
+		start_threads(threads, philosophers, args[0]);
+		start_death_checker_threads(threads, philosophers, args[0]);
+		join_threads(threads, args[0]);
+		destroy_forks(death_checker_locks, args[0]);
+	}
 	destroy_forks(forks, args[0]);
-	destroy_forks(death_checker_locks, args[0]);
 	deallocate(forks, death_checker_locks, philosophers, threads);
 	return (0);
 }
